feat(cli): verify that template category is the correct one
